refactor: elimina 4 duplicazioni CLI, sposta sql_dry_run in core
- dump_cfg_section → ensure_dict (common.py, -20)
- _quoted_identifier → q_ident (sql_dry_run.py, -3)
- sql_dry_run.py → core/sql_validation.py (+backward compat)
- _batch_helpers.py rimosso (orfano, -79)
- _run_pipeline/_execute_pipeline separati (cmd_run.py -292)
- ADR-003 superseded

Totale: -164 nette, 12 file, 1234 test
